requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
duckduckgo-search>=6.3.0
python-dotenv>=1.0.0
requests>=2.32.0
numpy>=1.26.0
